css/css-flexbox/stretch-table-child.html

<!DOCTYPE html>
<html>
<head>
<title>CSS Flexbox: Stretch a table</title>
<link rel="help" href="https://drafts.csswg.org/css-flexbox/#algo-stretch">
<link rel="help" href="https://drafts.csswg.org/css-flexbox/#flex-direction-property">
<meta name="assert" content="This test ensures that flexbox stretches children which has a table, in 'column' direction and 'flex: 0 0 auto'.">
<style>
.container {
    display: flex;
    flex-direction: column;
    height: 100px;
    width: 50px;
}
.first {
    flex: 1 1 auto;
    background-color: blue;
}
.test {
    flex: 0 0 auto;
    background-color: green;
    display: flex;
}
td {
    padding: 23px;
}
</style>
</head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<body onload="checkLayout('.test')">
<div id=log></div>
<div class="container">
    <div class="first">
    </div>
    <div class="test" data-expected-height=50>
        <table>
            <tr><td></td></tr>
        </table>
    </div>
</div>
</body>
</html>
